Pinyin and Baidu translate of tool
①BaiduTranslate:
It can translate long texts, and the general translation API supports the translation of 28 languages, covering Chinese, English, Japanese, Korean, Western, French, Thai, Arab, Russian, Portuguese, German, Italian, Dutch, Finnish, Danish and so on. It also supports the language detection of 28 languages.
②ZhToPinyin:
The basic principle of implementation is to find the ASK codes corresponding to Chinese characters, and then match the ASK codes with the pre-established Pinyin ASK codes dictionary to find the corresponding Pinyin letters. So whether we can get the results we want to convert accurately depends entirely on whether the ASK codes dictionary we built is complete and powerful.Comparing a lot of classes and methods found on the internet, after sorting out, it is still the most perfect ASK code dictionary of this kind, and many uncommon words can also be converted out.